+++ config/modules.php
@@ -35,4 +35,5 @@ return [
     TT\Modules\Pdp\PdpModule::class                       => true,
     TT\Modules\TeamDevelopment\TeamDevelopmentModule::class => true,
     TT\Modules\Trials\TrialsModule::class                 => true,
+    TT\Modules\Wizards\WizardsModule::class               => true,
 ];